+++ src/frontend/components/Forside/Forside.tsx
@@ -1,11 +1,9 @@
-import React, { useState, useEffect } from 'react';
+import React, { useEffect } from 'react';
 
 import { FormattedMessage, useIntl } from 'react-intl';
-import { useHistory } from 'react-router';
 import styled from 'styled-components/macro';
 
 import navFarger from 'nav-frontend-core';
-import KnappBase from 'nav-frontend-knapper';
 import Lenke from 'nav-frontend-lenker';
 import { Sidetittel, Normaltekst } from 'nav-frontend-typografi';
 
@@ -14,9 +12,8 @@ import { RessursStatus } from '@navikt/familie-typer';
 
 import VeilederSnakkeboble from '../../assets/VeilederSnakkeboble';
 import { useApp } from '../../context/AppContext';
-import { StegRoutes } from '../../routing/Routes';
 import Informasjonsbolk from '../Felleskomponenter/Informasjonsbolk/Informasjonsbolk';
-import BekreftelsePanel from './BekreftelsePanel';
+import BekreftelseOgStartSoknad from './BekreftelseOgStartSoknad';
 
 const panelBredde = '524px';
 const tablet = '959px';
@@ -51,23 +48,8 @@ const StyledLenke = styled(Lenke)`
     display: inline-block;
 `;
 
-const StyledKnappBase = styled(KnappBase)`
-    margin: 2.3rem auto 0 auto;
-`;
-
-export enum BekreftelseStatus {
-    NORMAL = 'NORMAL',
-    BEKREFTET = 'BEKREFTET',
-    FEIL = 'FEIL',
-}
-
 const Forside: React.FC = () => {
-    const [bekreftelseStatus, settBekreftelseStatus] = useState<BekreftelseStatus>(
-        BekreftelseStatus.NORMAL
-    );
-
     const { formatMessage } = useIntl();
-    const history = useHistory();
 
     const { sluttbruker, nullstillSøknadsobjekt, settUtfyltSteg } = useApp();
 
@@ -77,20 +59,6 @@ const Forside: React.FC = () => {
         nullstillSøknadsobjekt(sluttbruker);
         settUtfyltSteg(0);
     }, []);
-
-    const startSøknadOnClick = () => {
-        bekreftelseStatus === BekreftelseStatus.BEKREFTET
-            ? history.push(Object.values(StegRoutes)[0].path)
-            : settBekreftelseStatus(BekreftelseStatus.FEIL);
-    };
-
-    const bekreftelseOnChange = () => {
-        settBekreftelseStatus(prevState =>
-            prevState !== BekreftelseStatus.BEKREFTET
-                ? BekreftelseStatus.BEKREFTET
-                : BekreftelseStatus.NORMAL
-        );
-    };
 
     return (
         <ForsideContainer>
@@ -169,18 +137,7 @@ const Forside: React.FC = () => {
                 </Normaltekst>
             </Informasjonsbolk>
 
-            <BekreftelsePanel
-                navn={navn}
-                bekreftelseStatus={bekreftelseStatus}
-                onChange={bekreftelseOnChange}
-            />
-
-            <StyledKnappBase
-                type={bekreftelseStatus === BekreftelseStatus.BEKREFTET ? 'hoved' : 'standard'}
-                onClick={startSøknadOnClick}
-            >
-                Start søknaden
-            </StyledKnappBase>
+            <BekreftelseOgStartSoknad navn={navn} />
         </ForsideContainer>
     );
 };
